requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.6
asyncio==3.4.3
pandas==2.2.1
matplotlib==3.8.0
plotly==5.17.0
APScheduler==3.10.4
python-dotenv==1.0.0
lxml==4.9.3
gunicorn==21.2.0
